Move ZeroMatrix and GenericZeroMatrix
