+++ sympy/matrices/expressions/tests/test_permutation.py
@@ -4,7 +4,8 @@ from sympy.matrices import Matrix
 from sympy.matrices.expressions import \
     MatMul, BlockDiagMatrix, Determinant, Inverse
 from sympy.matrices.expressions.matexpr import \
-    MatrixSymbol, Identity, ZeroMatrix, OneMatrix
+    MatrixSymbol, Identity, OneMatrix
+from sympy.matrices.expressions.special import ZeroMatrix
 from sympy.matrices.expressions.permutation import \
     MatrixPermute, PermutationMatrix
 from sympy.testing.pytest import raises
